Support parsing annotations with dotted values
This was tripping up on this line from QtGui.QPainter when running on an
updated PyQt (6.3.1):

    def drawConvexPolygon(self, points: PyQt6.sip.array[QtCore.QPointF]) -> None: ...

Here PyQt6.sip.array is Attribute(Attribute(Name)) and we can supported
arbitrarily dotted names by recursing.
Also there is an attribute in the slice too which apparently wasn't
supported before.

I'm not sure what test case this should/could have. Importing QtGui and
calling the function with the right signature?