Remove reference to `pip install PyQt6-stubs`
When ~reading~only skimming the install instructions one comes around the first suggestion of `pip install PyQt6-stubs`, which we run and we get an error. _Then_ we notice the greyed out warning, curse a little, and follow the only useful instructions. It's a small paper cut, but also a small fix until this lands on PyPI.